Add template test support
