+++ package.json
@@ -38,6 +38,7 @@
     "@babel/parser": "^7.6.0",
     "chalk": "^2.4.2",
     "common-tags": "^1.8.0",
+    "ember-template-recast": "^4.1.4",
     "execa": "^2.0.4",
     "fs-extra": "^8.1.0",
     "globby": "^10.0.1",
