+++ src/test-support.js
@@ -1,79 +1,14 @@
 'use strict';
 
-/* global it, describe, beforeEach, afterEach */
-
-const { runInlineTest } = require('jscodeshift/dist/testUtils');
-const fs = require('fs-extra');
-const path = require('path');
-const globby = require('globby');
-
-function transformDetails(options) {
-  let root = process.cwd() + `/transforms/${options.name}/`;
-
-  return {
-    name: options.name,
-    root,
-    transformPath: root + 'index',
-    fixtureDir: root + '__testfixtures__/',
-  };
-}
-
-function jscodeshiftTest(options) {
-  let details = transformDetails(options);
-
-  let transform = require(details.transformPath);
-
-  describe(details.name, function() {
-    globby
-      .sync('**/*.input.*', {
-        cwd: details.fixtureDir,
-        absolute: true,
-      })
-      .map(entry => entry.slice(entry.indexOf('__testfixtures__') + '__testfixtures__'.length + 1))
-      .forEach(filename => {
-        let extension = path.extname(filename);
-        let testName = filename.replace(`.input${extension}`, '');
-        let testInputPath = path.join(details.fixtureDir, `${testName}${extension}`);
-        let inputPath = path.join(details.fixtureDir, `${testName}.input${extension}`);
-        let outputPath = path.join(details.fixtureDir, `${testName}.output${extension}`);
-        let optionsPath = path.join(details.fixtureDir, `${testName}.options.json`);
-        let options = fs.pathExistsSync(optionsPath) ? fs.readFileSync(optionsPath) : '{}';
-
-        describe(testName, function() {
-          beforeEach(function() {
-            process.env.CODEMOD_CLI_ARGS = options;
-          });
-
-          afterEach(function() {
-            process.env.CODEMOD_CLI_ARGS = '';
-          });
-
-          it('transforms correctly', function() {
-            runInlineTest(
-              transform,
-              {},
-              { path: testInputPath, source: fs.readFileSync(inputPath, 'utf8') },
-              fs.readFileSync(outputPath, 'utf8')
-            );
-          });
-
-          it('is idempotent', function() {
-            runInlineTest(
-              transform,
-              {},
-              { path: testInputPath, source: fs.readFileSync(outputPath, 'utf8') },
-              fs.readFileSync(outputPath, 'utf8')
-            );
-          });
-        });
-      });
-  });
-}
+const jscodeshiftTest = require('./test-support/jscodeshift');
+const templateTest = require('./test-support/template');
 
 function runTransformTest(options) {
   switch (options.type) {
     case 'jscodeshift':
       return jscodeshiftTest(options);
+    case 'template':
+      return templateTest(options);
   }
 }
