+++ src/test-support/utils.js
@@ -1,0 +1,16 @@
+'use strict';
+
+function transformDetails(options) {
+  let root = process.cwd() + `/transforms/${options.name}/`;
+
+  return {
+    name: options.name,
+    root,
+    transformPath: root + 'index',
+    fixtureDir: root + '__testfixtures__/',
+  };
+}
+
+module.exports = {
+  transformDetails,
+};
